requirements.txt
sanic==23.12.1
sanic-cors==2.2.0
sqlalchemy==2.0.43
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.1
PyJWT==2.8.0
bcrypt==4.1.2
aiofiles==24.1.0
websockets==12.0
python-multipart==0.0.6
email-validator==2.1.1
pydantic==2.5.3
python-dotenv==1.0.0
uvloop==0.19.0
aiosmtplib==3.0.1
greenlet==3.2.4
